+++ components/icu/README.md
@@ -120,12 +120,9 @@ are on track to be eventually stabilized into this crate.
 
 [CLDR]: http://cldr.unicode.org/
 [`DataProvider`]: icu_provider::DataProvider
-[`DataPayload`]: icu_provider::DataPayload
 [`FsDataProvider`]: https://docs.rs/icu_provider_fs/latest/icu_provider_fs/struct.FsDataProvider.html
 [`BlobDataProvider`]: https://docs.rs/icu_provider_blob/latest/icu_provider_blob/struct.BlobDataProvider.html
 [`icu_provider_adapters`]: https://docs.rs/icu_provider_adapters/latest/icu_provider_adapters/
-[`icu_datagen`]: https://docs.rs/icu_datagen/latest/icu_datagen/
-[`icu4x-datagen`]: https://docs.rs/icu_datagen/latest/icu_datagen/
 [data management tutorial]: https://github.com/unicode-org/icu4x/blob/main/tutorials/data_provider.md#loading-additional-data-at-runtime
 
 <!-- cargo-rdme end -->
