+++ provider/blob/src/export/mod.rs
@@ -4,14 +4,14 @@
 
 //! Data exporter that creates a binary blob for use with [`BlobDataProvider`](crate::BlobDataProvider).
 //!
-//! This module can be used as a target for the `icu_datagen` crate.
+//! This module can be used as a target for the `icu_provider_export` crate.
 //!
 //! See our [datagen tutorial](https://github.com/unicode-org/icu4x/blob/main/tutorials/data_management.md) for more information about different data providers.
 //!
 //! # Examples
 //!
 //! ```
-//! use icu_datagen::prelude::*;
+//! use icu_provider_export::prelude::*;
 //! use icu_provider_blob::export::*;
 //!
 //! let mut blob: Vec<u8> = Vec::new();
@@ -20,7 +20,7 @@
 //! let mut exporter = BlobExporter::new_v2_with_sink(Box::new(&mut blob));
 //!
 //! // Export something. Make sure to use the same fallback data at runtime!
-//! DatagenDriver::new([LocaleFamily::FULL], FallbackOptions::maximal_deduplication(), LocaleFallbacker::new().static_to_owned())
+//! ExportDriver::new([LocaleFamily::FULL], DeduplicationStrategy::Maximal.into(), LocaleFallbacker::new().static_to_owned())
 //!     .export(&icu_provider::hello_world::HelloWorldProvider, exporter)
 //!     .unwrap();
 //!
