+++ provider/export/src/export_impl.rs
@@ -2,491 +2,32 @@
 // called LICENSE at the top level of the ICU4X source tree
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
-use crate::rayon_prelude::*;
-use displaydoc::Display;
+use crate::{DeduplicationStrategy, ExportDriver, LocaleFamilyAnnotations};
 use icu_locale::extensions::unicode::key;
 use icu_locale::fallback::LocaleFallbackIterator;
 use icu_locale::LanguageIdentifier;
 use icu_locale::LocaleFallbacker;
-use icu_locale::ParseError;
 use icu_provider::export::*;
 use icu_provider::prelude::*;
 use std::collections::HashMap;
 use std::collections::HashSet;
 use std::fmt;
-use std::hash::Hash;
-use std::str::FromStr;
 use std::time::Duration;
 use std::time::Instant;
 use writeable::Writeable;
 
-/// Options bag configuring locale inclusion and behavior when runtime fallback is disabled.
-#[non_exhaustive]
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Default)]
-pub struct NoFallbackOptions {}
-
-/// Choices for determining the deduplication of locales for exported data payloads.
-///
-/// Deduplication affects the lookup table from locales to data payloads. If a child locale
-/// points to the same payload as its parent locale, then the child locale can be removed from
-/// the lookup table. Therefore, all deduplication strategies guarantee that data requests for
-/// selected locales will succeed so long as fallback is enabled at runtime (either internally
-/// or externally). They also do not impact which _payloads_ are included: only the lookup table.
-///
-/// Comparison of the deduplication strategies:
-///
-/// | Name | Data file size | Supported locale queries? | Needs runtime fallback? |
-/// |---|---|---|---|
-/// | [`Maximal`] | Smallest | No | Yes |
-/// | `RetainBaseLanguages` (TODO(#58): coming soon) | Small | Yes | Yes |
-/// | [`None`] | Medium/Small | Yes | No |
-///
-/// [`Maximal`]: DeduplicationStrategy::Maximal
-/// [`None`]: DeduplicationStrategy::None
-#[non_exhaustive]
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
-pub enum DeduplicationStrategy {
-    /// Removes from the lookup table any locale whose parent maps to the same data.
-    Maximal,
-    /// Removes from the lookup table any locale whose parent maps to the same data, except if
-    /// the parent is `und`.
-    RetainBaseLanguages,
-    /// Keeps all selected locales in the lookup table.
-    None,
-}
-
-/// Inner fields of a [`LocaleFamily`].
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
-pub(crate) struct LocaleFamilyAnnotations {
-    include_ancestors: bool,
-    include_descendants: bool,
-}
-
-impl LocaleFamilyAnnotations {
-    #[inline]
-    pub(crate) const fn with_descendants() -> Self {
-        Self {
-            include_ancestors: true,
-            include_descendants: true,
-        }
-    }
-
-    #[inline]
-    pub(crate) const fn without_descendants() -> Self {
-        Self {
-            include_ancestors: true,
-            include_descendants: false,
-        }
-    }
-
-    #[inline]
-    pub(crate) const fn without_ancestors() -> Self {
-        Self {
-            include_ancestors: false,
-            include_descendants: true,
-        }
-    }
-
-    #[inline]
-    pub(crate) const fn single() -> Self {
-        Self {
-            include_ancestors: false,
-            include_descendants: false,
-        }
+#[cfg(not(feature = "rayon"))]
+trait IntoParallelIterator: IntoIterator + Sized {
+    fn into_par_iter(self) -> <Self as IntoIterator>::IntoIter {
+        self.into_iter()
     }
 }
+#[cfg(not(feature = "rayon"))]
+impl<T: IntoIterator> IntoParallelIterator for T {}
+#[cfg(feature = "rayon")]
+use rayon::prelude::*;
 
-impl Writeable for LocaleFamilyAnnotations {
-    fn write_to<W: core::fmt::Write + ?Sized>(&self, sink: &mut W) -> core::fmt::Result {
-        match (self.include_ancestors, self.include_descendants) {
-            (true, true) => Ok(()),
-            (true, false) => sink.write_char('^'),
-            (false, true) => sink.write_char('%'),
-            (false, false) => sink.write_char('@'),
-        }
-    }
-
-    fn writeable_length_hint(&self) -> writeable::LengthHint {
-        writeable::LengthHint::exact(match (self.include_ancestors, self.include_descendants) {
-            (true, true) => 0,
-            _ => 1,
-        })
-    }
-}
-
-/// A family of locales to export.
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub struct LocaleFamily {
-    langid: Option<LanguageIdentifier>,
-    annotations: LocaleFamilyAnnotations,
-}
-
-impl LocaleFamily {
-    /// The family containing all ancestors and descendants of the selected locale.
-    ///
-    /// This is the recommended family type since it reflects regional preferences.
-    ///
-    /// For example, the family `::with_descendants("en-001")` contains:
-    ///
-    /// - Self: "en-001"
-    /// - Ancestors: "und", "en"
-    /// - Descendants: "en-GB", "en-ZA", ...
-    ///
-    /// Stylized on the CLI as: "en-US"
-    ///
-    /// The `und` locale is treated specially and behaves like `::single("und")`.
-    pub const fn with_descendants(langid: LanguageIdentifier) -> Self {
-        let annotations = if langid.is_empty() {
-            LocaleFamilyAnnotations::single()
-        } else {
-            LocaleFamilyAnnotations::with_descendants()
-        };
-
-        Self {
-            langid: Some(langid),
-            annotations,
-        }
-    }
-
-    /// The family containing all ancestors of the selected locale.
-    ///
-    /// This family type does not include regional variants unless the selected locale is itself
-    /// a regional variant.
-    ///
-    /// For example, the family `::without_descendants("en-001")` contains:
-    ///
-    /// - Self: "en-001"
-    /// - Ancestors: "und", "en"
-    ///
-    /// Stylized on the CLI as: "^en-US"
-    ///
-    /// The `und` locale is treated specially and behaves like `::single("und")`.
-    pub const fn without_descendants(langid: LanguageIdentifier) -> Self {
-        let annotations = if langid.is_empty() {
-            LocaleFamilyAnnotations::single()
-        } else {
-            LocaleFamilyAnnotations::without_descendants()
-        };
-        Self {
-            langid: Some(langid),
-            annotations,
-        }
-    }
-
-    /// The family containing all descendants of the selected locale.
-    ///
-    /// This family may be useful if the root locale is not desired.
-    ///
-    /// For example, the family `::without_ancestors("en-001")` contains:
-    ///
-    /// - Self: "en-001"
-    /// - Descendants: "en-GB", "en-ZA", ...
-    ///
-    /// but it does _not_ contain the ancestors "en" and "und".
-    ///
-    /// Stylized on the CLI as: "%en-US"
-    ///
-    /// The `und` locale is treated specially and behaves like `::single("und")`.
-    pub const fn without_ancestors(langid: LanguageIdentifier) -> Self {
-        let annotations = if langid.is_empty() {
-            LocaleFamilyAnnotations::single()
-        } else {
-            LocaleFamilyAnnotations::without_ancestors()
-        };
-        Self {
-            langid: Some(langid),
-            annotations,
-        }
-    }
-
-    /// The family containing only the selected locale.
-    ///
-    /// For example, the family `::single("en-001")` contains only "en-001".
-    ///
-    /// Stylized on the CLI as: "@en-US"
-    pub const fn single(langid: LanguageIdentifier) -> Self {
-        Self {
-            langid: Some(langid),
-            annotations: LocaleFamilyAnnotations::single(),
-        }
-    }
-
-    /// The family containing all locales.
-    ///
-    /// Stylized on the CLI as: "full"
-    pub const FULL: Self = Self {
-        langid: None,
-        annotations: LocaleFamilyAnnotations {
-            include_ancestors: false,
-            include_descendants: true,
-        },
-    };
-}
-
-impl Writeable for LocaleFamily {
-    fn write_to<W: core::fmt::Write + ?Sized>(&self, sink: &mut W) -> core::fmt::Result {
-        if let Some(langid) = self.langid.as_ref() {
-            self.annotations.write_to(sink)?;
-            langid.write_to(sink)
-        } else {
-            sink.write_str("full")
-        }
-    }
-
-    fn writeable_length_hint(&self) -> writeable::LengthHint {
-        if let Some(langid) = self.langid.as_ref() {
-            self.annotations.writeable_length_hint() + langid.writeable_length_hint()
-        } else {
-            writeable::LengthHint::exact(4)
-        }
-    }
-}
-
-writeable::impl_display_with_writeable!(LocaleFamily);
-
-/// An error while parsing a [`LocaleFamily`].
-#[derive(Debug, Copy, Clone, PartialEq, Display)]
-#[non_exhaustive]
-pub enum LocaleFamilyParseError {
-    /// An error bubbled up from parsing a [`LanguageIdentifier`].
-    #[displaydoc("{0}")]
-    LanguageIdentifier(ParseError),
-    /// Some other error specific to parsing the family, such as an invalid lead byte.
-    #[displaydoc("Invalid locale family")]
-    InvalidFamily,
-}
-
-impl From<ParseError> for LocaleFamilyParseError {
-    fn from(err: ParseError) -> Self {
-        Self::LanguageIdentifier(err)
-    }
-}
-
-impl std::error::Error for LocaleFamilyParseError {}
-
-impl FromStr for LocaleFamily {
-    type Err = LocaleFamilyParseError;
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        if s == "full" {
-            return Ok(Self::FULL);
-        }
-        let (first, remainder) = s
-            .as_bytes()
-            .split_first()
-            .ok_or(LocaleFamilyParseError::InvalidFamily)?;
-        match first {
-            b'^' => Ok(Self {
-                langid: Some(LanguageIdentifier::try_from_utf8(remainder)?),
-                annotations: LocaleFamilyAnnotations::without_descendants(),
-            }),
-            b'%' => Ok(Self {
-                langid: Some(LanguageIdentifier::try_from_utf8(remainder)?),
-                annotations: LocaleFamilyAnnotations::without_ancestors(),
-            }),
-            b'@' => Ok(Self {
-                langid: Some(LanguageIdentifier::try_from_utf8(remainder)?),
-                annotations: LocaleFamilyAnnotations::single(),
-            }),
-            b if b.is_ascii_alphanumeric() => Ok(Self {
-                langid: Some(s.parse()?),
-                annotations: LocaleFamilyAnnotations::with_descendants(),
-            }),
-            _ => Err(LocaleFamilyParseError::InvalidFamily),
-        }
-    }
-}
-
-#[test]
-fn test_locale_family_parsing() {
-    let valid_families = ["und", "de-CH", "^es", "@pt-BR", "%en-001", "full"];
-    let invalid_families = ["invalid", "@invalid", "-foo", "@full", "full-001"];
-    for family_str in valid_families {
-        let family = family_str.parse::<LocaleFamily>().unwrap();
-        let family_to_str = family.to_string();
-        assert_eq!(family_str, family_to_str);
-    }
-    for family_str in invalid_families {
-        assert!(family_str.parse::<LocaleFamily>().is_err());
-    }
-}
-
-/// Options bag configuring locale inclusion and behavior when runtime fallback is enabled.
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
-#[non_exhaustive]
-pub struct FallbackOptions {
-    /// The aggressiveness of deduplication of data payloads.
-    pub deduplication_strategy: DeduplicationStrategy,
-}
-
-impl FallbackOptions {
-    /// Creates a [`FallbackOptions`] with [`DeduplicationStrategy::None`]
-    pub fn no_deduplication() -> Self {
-        Self {
-            deduplication_strategy: DeduplicationStrategy::None,
-        }
-    }
-
-    /// Creates a [`FallbackOptions`] with [`DeduplicationStrategy::Maximal`]
-    pub fn maximal_deduplication() -> Self {
-        Self {
-            deduplication_strategy: DeduplicationStrategy::Maximal,
-        }
-    }
-
-    /// Creates a [`FallbackOptions`] with [`DeduplicationStrategy::RetainBaseLanguages`]
-    pub fn retain_base_languages_deduplication() -> Self {
-        Self {
-            deduplication_strategy: DeduplicationStrategy::RetainBaseLanguages,
-        }
-    }
-}
-
-/// Configuration for a data export operation.
-///
-/// Note that this only configures *which data* is exported. The input provider, usually
-/// `DatagenProvider`, might expose more options about the data itself.
-///
-/// # Examples
-///
-/// ```no_run
-/// use icu_datagen::blob_exporter::*;
-/// use icu_datagen::prelude::*;
-/// use icu_datagen_bikeshed::DatagenProvider;
-///
-/// let provider = DatagenProvider::new_latest_tested();
-///
-/// DatagenDriver::new([LocaleFamily::FULL], FallbackOptions::no_deduplication(), LocaleFallbacker::try_new_unstable(&provider).unwrap())
-///     .with_markers([icu::list::provider::AndListV2Marker::INFO])
-///     .export(
-///         &provider,
-///         BlobExporter::new_with_sink(Box::new(&mut Vec::new())),
-///     )
-///     .unwrap();
-/// ```
-#[derive(Debug, Clone)]
-pub struct DatagenDriver {
-    markers: Option<HashSet<DataMarkerInfo>>,
-    requested_families: HashMap<LanguageIdentifier, LocaleFamilyAnnotations>,
-    fallbacker: LocaleFallbacker,
-    include_full: bool,
-    deduplication_strategy: DeduplicationStrategy,
-    additional_collations: HashSet<String>,
-    segmenter_models: Vec<String>,
-}
-
-impl DatagenDriver {
-    /// Creates a [`DatagenDriver`].
-    ///
-    /// The fallbacker is used to resolve locale families, and to dedpulicate data if requested.
-    /// Make sure to use the same fallback data when loading from the provider at runtime.
-    /// Commonly, you will export the fallback markers, in which case you should construct
-    /// your fallbacker with the source provider (i.e. [`LocaleFallbacker::try_new_unstable`]).
-    pub fn new(
-        locales: impl IntoIterator<Item = LocaleFamily>,
-        options: FallbackOptions,
-        fallbacker: LocaleFallbacker,
-    ) -> Self {
-        let mut include_full = false;
-        Self {
-            markers: Default::default(),
-            requested_families: locales
-                .into_iter()
-                .filter_map(|family| {
-                    Some((
-                        family.langid.or_else(|| {
-                            // Full locale family: set the bit instead of adding to the set
-                            debug_assert_eq!(family.annotations, LocaleFamily::FULL.annotations);
-                            include_full = true;
-                            None
-                        })?,
-                        family.annotations,
-                    ))
-                })
-                .collect(),
-            include_full,
-            fallbacker,
-            deduplication_strategy: options.deduplication_strategy,
-            additional_collations: Default::default(),
-            segmenter_models: Default::default(),
-        }
-        .with_recommended_segmenter_models()
-    }
-
-    /// Sets this driver to generate the given data markers.
-    ///
-    /// If this is not called, all markers supported by the provider will be exported.
-    pub fn with_markers(self, markers: impl IntoIterator<Item = DataMarkerInfo>) -> Self {
-        Self {
-            markers: Some(markers.into_iter().collect()),
-            ..self
-        }
-    }
-
-    /// This option is only relevant if using `icu::collator`.
-    ///
-    /// By default, the collations `big5han`, `gb2312`, and those starting with `search`
-    /// are excluded. This method can be used to reennable them.
-    ///
-    /// The special string `"search*"` causes all search collation tables to be included.
-    pub fn with_additional_collations(
-        self,
-        additional_collations: impl IntoIterator<Item = String>,
-    ) -> Self {
-        Self {
-            additional_collations: additional_collations.into_iter().collect(),
-            ..self
-        }
-    }
-
-    /// This option is only relevant if using `icu::segmenter`.
-    ///
-    /// Sets this driver to generate the recommended segmentation models, to the extent required by the
-    /// chosen data markers.
-    pub fn with_recommended_segmenter_models(self) -> Self {
-        self.with_segmenter_models([
-            "cjdict".into(),
-            "burmesedict".into(),
-            "khmerdict".into(),
-            "laodict".into(),
-            "thaidict".into(),
-            "Burmese_codepoints_exclusive_model4_heavy".into(),
-            "Khmer_codepoints_exclusive_model4_heavy".into(),
-            "Lao_codepoints_exclusive_model4_heavy".into(),
-            "Thai_codepoints_exclusive_model4_heavy".into(),
-        ])
-    }
-
-    /// This option is only relevant if using `icu::segmenter`.
-    ///
-    /// Sets this driver to generate the given segmentation models, to the extent required by the
-    /// chosen data markers.
-    ///
-    /// The currently supported dictionary models are
-    /// * `cjdict`
-    /// * `burmesedict`
-    /// * `khmerdict`
-    /// * `laodict`
-    /// * `thaidict`
-    ///
-    /// The currently supported LSTM models are
-    /// * `Burmese_codepoints_exclusive_model4_heavy`
-    /// * `Khmer_codepoints_exclusive_model4_heavy`
-    /// * `Lao_codepoints_exclusive_model4_heavy`
-    /// * `Thai_codepoints_exclusive_model4_heavy`
-    ///
-    /// If a model is not included, the resulting line or word segmenter will apply rule-based
-    /// segmentation when encountering text in a script that requires the model, which will be
-    /// incorrect.
-    ///
-    /// If multiple models for the same language and segmentation type (dictionary/LSTM) are
-    /// listed, the first one will be used.
-    pub fn with_segmenter_models(self, models: impl IntoIterator<Item = String>) -> Self {
-        Self {
-            segmenter_models: models.into_iter().collect(),
-            ..self
-        }
-    }
-
+impl ExportDriver {
     /// Exports data from the given provider to the given exporter.
     ///
     /// See
@@ -1131,36 +672,4 @@ fn test_collation_filtering() {
             .collect::<BTreeSet<_>>();
         assert_eq!(resolved_locales, expected_locales, "{cas:?}");
     }
-}
-
-/// Test that the last option with multiple conflicting families wins.
-#[test]
-fn test_family_precedence() {
-    let driver = DatagenDriver::new(
-        [
-            "en".parse().unwrap(),
-            "%en".parse().unwrap(),
-            "@en".parse().unwrap(),
-            "%zh-TW".parse().unwrap(),
-            "^zh-TW".parse().unwrap(),
-        ],
-        FallbackOptions::no_deduplication(),
-        LocaleFallbacker::new_without_data(),
-    );
-
-    assert_eq!(
-        driver.requested_families,
-        [
-            (
-                icu::locale::langid!("en"),
-                LocaleFamilyAnnotations::single()
-            ),
-            (
-                icu::locale::langid!("zh-TW"),
-                LocaleFamilyAnnotations::without_descendants()
-            ),
-        ]
-        .into_iter()
-        .collect::<HashMap<_, _>>()
-    );
 }
