+++ provider/fs/Cargo.toml
@@ -36,7 +36,7 @@ serde_json = { workspace = true, features = ["std"], optional = true }
 icu_benchmark_macros = { path = "../../tools/benchmark/macros" }
 icu_locale_core = { path = "../../components/locale_core", default-features = false, features = ["serde"] }
 icu_provider = { path = "../../provider/core", default-features = false, features = ["deserialize_json", "deserialize_bincode_1", "deserialize_postcard_1", "export"] }
-icu_datagen = { path = "../../provider/datagen", default-features = false }
+icu_provider_export = { path = "../../provider/export", default-features = false }
 writeable = { path = "../../utils/writeable" }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dev-dependencies]
