+++ tools/md-tests/Cargo.toml
@@ -10,7 +10,7 @@ edition = "2021"
 
 [dev-dependencies]
 icu = { workspace = true, features = ["compiled_data", "serde"] }
-icu_datagen = { workspace = true }
+icu_provider_export = { workspace = true }
 icu_provider = { workspace = true, features = ["deserialize_json"] }
 icu_provider_adapters = { workspace = true, features = ["serde"] }
 icu_provider_blob = { workspace = true }
