+++ tutorials/.cargo/config.toml
@@ -35,7 +35,7 @@ icu_freertos = { path = "../ffi/freertos" }
 icu_harfbuzz = { path = "../ffi/harfbuzz" }
 
 # Provider
-icu_datagen = { path = "../provider/datagen" }
+icu_provider_export = { path = "../provider/export" }
 icu_datagen_bikeshed = { path = "../provider/bikeshed" }
 icu_provider = { path = "../provider/core" }
 icu_provider_macros = { path = "../provider/core/macros" }
